feat: moved setEndDate and setDaysArray methods into the Week model instead of a helpers file
